Reorganizacion estructura de proyecto: carpetas raíz de app a src/app y actualizacion de importaciones
