+++ src/app/app.config.ts
@@ -1,7 +1,7 @@
 import { ApplicationConfig, provideBrowserGlobalErrorListeners, importProvidersFrom } from '@angular/core';
 import { provideRouter } from '@angular/router';
 import { provideHttpClient, withInterceptors } from '@angular/common/http';
-import { jwtInterceptor } from '../interceptors/jwt.interceptor';
+import { jwtInterceptor } from './interceptors/jwt.interceptor';
 import { routes } from './app.routes';
 
 export const appConfig: ApplicationConfig = {
