+++ src/app/services/student.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { LoginService } from '../../services/login.service';
+import { LoginService } from './login.service';
 
 export interface StudentData {
   nombre: string;
